requirements.txt
python-whois@0.7.3 